backend/main/phising login/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-dotenv==1.0.0